requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.30.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
FastAPI-SQLAlchemy==0.2.1
fonttools==4.59.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
kombu==5.5.4
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.6
motor==3.7.1
numpy==2.3.3
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pillow==11.3.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyodbc==5.2.0
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
redis==6.4.0
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
win32_setctime==1.2.0
